requirements.txt
fastapi
uvicorn
bcrypt
pyjwt
python-dotenv
pdfplumber
python-multipart
starlette
python-docx
requests
transformers
sentence-transformers
torch
httpx
<<<<<<< HEAD
nltk
=======
openai
nltk
pytest
>>>>>>> origin/feature/justin-a
